backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11,<4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135",
    "httpx>=0.28.1",
    "pydantic-settings>=2",
    "sqlalchemy[asyncio]>=2.0.48",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
